# docs/contracts/BASIC_CURRENT_DEFINITION_OPENAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DCP-BASIC-NO-AUDIT-VERSION-001 bounded API supplement",
    "version": "2026-10-09",
    "description": "Approved supplement to FINAL BASELINE v1.0.23. Only listed basic controls/process bindings supersede prior contracts. Historical evidence and regulated controls remain. Fully qualified schema IDs prevent Create/Update collisions without changing HTTP payload fields. DCP-EBR-TEMPLATE-NAME-001 adds required create templateName and controlled draft name editing; historical canonical content remains unchanged."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/api/v1/suppliers/{id}": {
      "get": {
        "tags": [
          "supplier-controller"
        ],
        "operationId": "get_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.hospital.mes.common.api.ApiResponseCom.fasterxml.jackson.databind.JsonNode"
                }
              }
            }
          }
        },
        "x-basic-maintenance-contract": "No audit append or revision/client-version precondition. Permission, org scope, validation, row locks and idempotency retained."
      },
      "put": {
        "tags": [
          "supplier-controller"
        ],
        "operationId": "update_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/com.hospital.mes.masterdata.application.SupplierCommands.Update"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.hospital.mes.common.api.ApiResponseCom.fasterxml.jackson.databind.JsonNode"
                }
              }
            }
          }
        },
        "x-basic-maintenance-contract": "No audit append or revision/client-version precondition. Permission, org scope, validation, row locks and idempotency retained."
      }
    },
    "/api/v1/qualifications/{id}": {
      "get": {
        "tags": [
          "qualification-controller"
        ],
        "operationId": "get_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.hospital.mes.common.api.ApiResponseCom.fasterxml.jackson.databind.JsonNode"
                }
              }
            }
          }
        },
        "x-basic-maintenance-contract": "No audit append or revision/client-version precondition. Permission, org scope, validation, row locks and idempotency retained."
      },
      "put": {
        "tags": [
          "qualification-controller"
        ],
        "operationId": "update_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/com.hospital.mes.masterdata.application.QualificationCommands.Update"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.hospital.mes.common.api.ApiResponseCom.fasterxml.jackson.databind.JsonNode"
                }
              }
            }
          }
        },
        "x-basic-maintenance-contract": "No audit append or revision/client-version precondition. Permission, org scope, validation, row locks and idempotency retained."
      }
    },
...
    "/api/v1/organizations/{id}": {
      "get": {
        "tags": [
          "organization-controller"
        ],
        "operationId": "get_5",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.hospital.mes.common.api.ApiResponseCom.fasterxml.jackson.databind.JsonNode"
                }
              }
            }
          }
        },
        "x-basic-maintenance-contract": "No audit append or revision/client-version precondition. Permission, org scope, validation, row locks and idempotency retained."
      },
      "put": {
        "tags": [
          "organization-controller"
        ],
        "operationId": "update_5",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/com.hospital.mes.masterdata.application.OrganizationCommands.Update"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.hospital.mes.common.api.ApiResponseCom.fasterxml.jackson.databind.JsonNode"
                }
              }
            }
          }
        },
        "x-basic-maintenance-contract": "No audit append or revision/client-version precondition. Permission, org scope, validation, row locks and idempotency retained."
      }
    },
...
    "/api/v1/materials/{id}/suppliers": {
      "get": {
        "tags": [
          "material-controller"
        ],
        "operationId": "suppliers",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.hospital.mes.common.api.ApiResponseCom.fasterxml.jackson.databind.JsonNode"
                }
              }
            }
          }
        },
        "x-basic-maintenance-contract": "No audit append or revision/client-version precondition. Permission, org scope, validation, row locks and idempotency retained."
      },
      "put": {
        "tags": [
          "material-controller"
        ],
        "operationId": "assign_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/com.hospital.mes.masterdata.application.SupplierCommands.Assign"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.hospital.mes.common.api.ApiResponseCom.fasterxml.jackson.databind.JsonNode"
                }
              }
            }
          }
        },
        "x-basic-maintenance-contract": "No audit append or revision/client-version precondition. Permission, org scope, validation, row locks and idempotency retained."
      }
    },
...
    "/api/v1/suppliers": {
      "get": {
        "tags": [
          "supplier-controller"
        ],
        "operationId": "list_3",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "arg3",
            "in": "query",
            "required": true,
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.hospital.mes.common.api.ApiResponseCom.hospital.mes.masterdata.application.ScopedStore.PageDataCom.fasterxml.jackson.databind.JsonNode"
                }
              }
            }
          }
        },
        "x-basic-maintenance-contract": "No audit append or revision/client-version precondition. Permission, org scope, validation, row locks and idempotency retained."
      },
      "post": {
        "tags": [
          "supplier-controller"
        ],
        "operationId": "create_2",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/com.hospital.mes.masterdata.application.SupplierCommands.Create"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.hospital.mes.common.api.ApiResponseCom.fasterxml.jackson.databind.JsonNode"
                }
              }
            }
          }
        },
        "x-basic-maintenance-contract": "No audit append or revision/client-version precondition. Permission, org scope, validation, row locks and idempotency retained."
      }
    },
    "/api/v1/qualifications": {
      "get": {
        "tags": [
          "qualification-controller"
        ],
        "operationId": "list_6",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "arg3",
            "in": "query",
            "required": true,
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.hospital.mes.common.api.ApiResponseCom.hospital.mes.masterdata.application.ScopedStore.PageDataCom.fasterxml.jackson.databind.JsonNode"
                }
              }
            }
          }
        },
        "x-basic-maintenance-contract": "No audit append or revision/client-version precondition. Permission, org scope, validation, row locks and idempotency retained."
      },
      "post": {
        "tags": [
          "qualification-controller"
        ],
        "operationId": "create_6",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/com.hospital.mes.masterdata.application.QualificationCommands.Create"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.hospital.mes.common.api.ApiResponseCom.fasterxml.jackson.databind.JsonNode"
                }
              }
            }
          }
        },
        "x-basic-maintenance-contract": "No audit append or revision/client-version precondition. Permission, org scope, validation, row locks and idempotency retained."
      }
    },
...
    "/api/v1/organizations": {
      "get": {
        "tags": [
          "organization-controller"
        ],
        "operationId": "list_8",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "arg3",
            "in": "query",
            "required": true,
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.hospital.mes.common.api.ApiResponseCom.hospital.mes.masterdata.application.ScopedStore.PageDataCom.fasterxml.jackson.databind.JsonNode"
                }
              }
            }
          }
        },
        "x-basic-maintenance-contract": "No audit append or revision/client-version precondition. Permission, org scope, validation, row locks and idempotency retained."
      },
      "post": {
        "tags": [
          "organization-controller"
        ],
        "operationId": "create_7",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/com.hospital.mes.masterdata.application.OrganizationCommands.Create"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.hospital.mes.common.api.ApiResponseCom.fasterxml.jackson.databind.JsonNode"
                }
              }
            }
          }
        },
        "x-basic-maintenance-contract": "No audit append or revision/client-version precondition. Permission, org scope, validation, row locks and idempotency retained."
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "com.hospital.mes.common.api.ApiResponseCom.hospital.mes.masterdata.application.ScopedStore.PageDataCom.fasterxml.jackson.databind.JsonNode": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/com.hospital.mes.masterdata.application.ScopedStore.PageDataCom.fasterxml.jackson.databind.JsonNode"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
...
      "com.hospital.mes.masterdata.application.SupplierCommands.Create": {
        "type": "object",
        "properties": {
          "supplierCode": {
            "type": "string"
          },
          "supplierName": {
            "type": "string"
          },
          "validTo": {
            "type": "string",
            "format": "date"
          }
        },
        "additionalProperties": false
      },
...
      "com.hospital.mes.masterdata.application.OrganizationCommands.Create": {
        "type": "object",
        "properties": {
          "parentId": {
            "type": "string"
          },
          "orgCode": {
            "type": "string"
          },
          "orgName": {
            "type": "string"
          },
          "orgType": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "com.hospital.mes.common.api.ApiResponseCom.fasterxml.jackson.databind.JsonNode": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/com.fasterxml.jackson.databind.JsonNode"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
      "com.fasterxml.jackson.databind.JsonNode": {},
      "com.hospital.mes.masterdata.application.SupplierCommands.Assign": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          },
          "suppliers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/com.hospital.mes.masterdata.application.SupplierCommands.Relationship"
            }
          }
        },
        "additionalProperties": false
      },
...
      "com.hospital.mes.masterdata.application.OrganizationCommands.Update": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "parentId": {
            "type": "string"
          },
          "orgName": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "com.hospital.mes.masterdata.application.QualificationCommands.Update": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "validFrom": {
            "type": "string",
            "format": "date"
          },
          "validTo": {
            "type": "string",
            "format": "date"
          }
        },
        "additionalProperties": false
      },
      "com.hospital.mes.masterdata.application.QualificationCommands.Create": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "qualificationCode": {
            "type": "string"
          },
          "validFrom": {
            "type": "string",
            "format": "date"
          },
          "validTo": {
            "type": "string",
            "format": "date"
          }
        },
        "additionalProperties": false
      },
      "com.hospital.mes.masterdata.application.SupplierCommands.Update": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "supplierName": {
            "type": "string"
          },
          "validTo": {
            "type": "string",
            "format": "date"
          }
        },
        "additionalProperties": false
      },
...
      "com.hospital.mes.masterdata.application.SupplierCommands.Relationship": {
        "type": "object",
        "properties": {
          "supplierId": {
            "type": "string"
          },
          "approved": {
            "type": "boolean"
          },
          "preferred": {
            "type": "boolean"
          },
          "validTo": {
            "type": "string",
            "format": "date"
          },
          "manufacturerName": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "com.hospital.mes.masterdata.application.ScopedStore.PageDataCom.fasterxml.jackson.databind.JsonNode": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/com.fasterxml.jackson.databind.JsonNode"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}